Remove local yfinance stub and add optional fallback
