+++ tests/test_fetch_sector.py
@@ -1,12 +1,16 @@
 import os
 import subprocess
 import sys
+from pathlib import Path
 
 
 def test_fetch_sector_aapl():
-    env = dict(**os.environ, SANDBOX_HAS_DATABASE='no')
+    env = dict(**os.environ, SANDBOX_HAS_DATABASE='no', USE_YFINANCE_STUB='1')
+    repo_root = Path(__file__).resolve().parents[1]
+    venv_python = repo_root / '.venv' / 'bin' / 'python'
+    executable = str(venv_python if venv_python.exists() else sys.executable)
     result = subprocess.run(
-        [sys.executable, 'fetch_sector.py', 'AAPL', '--dummy-run'],
+        [executable, 'fetch_sector.py', 'AAPL', '--dummy-run'],
         capture_output=True, text=True, env=env
     )
     print(result.stdout)
