--- yfinance/__init__.py
@@ -1,6 +1,0 @@
-class Ticker:
-    def __init__(self, ticker):
-        self.ticker = ticker
-    @property
-    def info(self):
-        return {"sector": "Technology"}
